refac: Decimal => UDecimal
Since as of now the "Decimal" type is only defined for positive ints and
operations with positive results it is more suitable to call the type
UDecimal and its context, conversion and floats-calculator accordingly
UContext, UConversion and UFloat.  Consequently the default context Dec
becomes the default context UDec.  This refactoring creates space for a
future Decimal type which supports also negative numbers.